requirements.txt
# NOx LGBM 모델용 requirements.txt
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
mlflow>=2.8.0
pyyaml>=6.0
awslambdaric>=2.0.0 